выполнено ТЗ 11

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -8,6 +8,7 @@ import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.repository.UserRepository;
 
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
@@ -22,11 +23,16 @@ public class UserService {
     }
 
     public User get(long id) {
+        log.info("GET /users");
         return userRepository.get(id);
     }
 
-    public Set<Long> getFriendsIds(long id) {
-        return userRepository.getFriendsIds(id);
+    public List<User> getFriends(long id) {
+        log.info("GET /users{id}");
+        if (userRepository.get(id) == null) {
+            throw new NotFoundException("Пользователь с данным ID не найден");
+        }
+        return userRepository.getFriends(id);
     }
 
     public User add(User user) {
@@ -51,16 +57,33 @@ public class UserService {
         return user;
     }
 
-    public void addFriend(long userId, long friendId) {
-        userRepository.addFriend(userId, friendId);
+    public List<User> addFriend(long userId, long friendId) {
+        log.info("PUT /users/{id}/friends/{friendId}");
+        if (userRepository.get(friendId) == null) {
+            throw new NotFoundException("Друг с таким ID не найден");
+        }
+        if (userRepository.get(userId) == null) {
+            throw new NotFoundException("Пользователь с таким ID не найден");
+        }
+        return userRepository.addFriend(userId, friendId);
     }
 
-    public Long removeFriend(long userId, long friendId) {
-        userRepository.removeFriend(userId, friendId);
-        return userRepository.getFriendById(userId, friendId);
+    public List<User> removeFriend(long userId, long friendId) {
+        log.info("DELETE /users/{id}/friends/{friendId}");
+        if (userRepository.get(userId) == null) {
+            throw new NotFoundException("Пользователь с указанным ID не найден");
+        }
+        if (userRepository.get(friendId) == null) {
+            throw new NotFoundException("Пользователь с указанным ID не найден");
+        }
+        if (!userRepository.getFriends(userId).contains(userRepository.get(friendId))) {
+            return userRepository.getFriends(userId);
+        }
+        return userRepository.removeFriend(userId, friendId);
     }
 
-    public Set<Long> getMutualFriends(long userId, long friendId) {
+    public List<User> getMutualFriends(long userId, long friendId) {
+        log.info("GET /users/{id}/friends/common/{otherId}");
         return userRepository.getMutualFriends(userId, friendId);
     }
 }
